Add Donation Entity (#27)
In this change, we add donations to the subgraph. For donations, we need to include the `data` field from the claim params in the events. Unfortunately, the test code for the mappings has a lot of duplication because it's AssemblyScript and not Typescript. I tried many ways to remove that duplication into functions but I couldn't find a way. The compiler doesn't allow me and I've already spent too many hours on it.

Extra: Fixed docker deployment.