Add Donations contract (#44)
This contract is the last step before the donated funds are sent to the charities. This contract will receive the donations claimed by the treasury and store an NFT for every donation. This contract will be deployed to Polygon, and the funds will have to be bridged by the backend. The backend will also be in charge of minting the NFTs for everyone. 

The variable `_donationsId` in the mint function is used for transparency and the backend from creating more NFTs than it's supposed to. The idea is for this variable to be the tx hash where the treasury claimed the funds on Ethereum. You can look up the transaction and make sure the backend behaves as expected. 

On the `donate` function, the `_to` param is provided by TheGivingBlock. Each charity will have more than one address once we have non-anonymous donations.