Partial Withdrawals (#93)
Partial withdrawals are necessary when users want to withdraw, but there isn't enough liquidity in the vault. This can happen for multiple reasons, namely:

1. The withdrawal is bigger than the non-investable percentage. For instance, if the vault invests 95% of the funds and the withdrawals, but the deposit represents 6%.
2. Someone has already withdrawn from the vault, and it has not yet rebalanced with the strategy.

A new external function is introduced, taking in a list of ids and a list of amounts. The idea is to delegate to the caller the responsibility of choosing the deposits to withdraw from.

Say Alice deposits 100 UST and assigns 30% of the yield to Bob and 70% to Carol. She may then want to do either of the following:

* do a partial withdrawal of 70 UST, all from Carol's share, leaving only Bob's share
* do a proportional withdrawal from each of them (70 * 0.7 from Carol and 70 * 0.3 from Bob)
* do some other combination of values that amounts to 70 UST

In this way, we can choose any of these approaches on the UI and iterate later on.

### Async withdrawals

This removes the need to implement async withdrawals, which are more complex (both technically and UX-wise).
If a user owns more than 5% of a vault, he would now be able to make a partial withdrawal of that 5% immediately instead of requesting an async withdrawal. This would also give us time to intervene if needed, should a bank run occur and allow us to increase the liquidity available in the vault to respond to all intended withdrawals.

### Force withdraws

The vault also has a force feature that allows users to withdraw at a loss (e.g., a black swan event that causes a negative yield). To avoid writing a mess of code, partial withdrawals are not allowed during loss scenarios, only full withdrawals with force.