Allow editing a Metavault (#117)
The concept of Metavault doesn't have a representation on the contracts. The way we link multiple Deposit is using a groupId. For the UI, deposits with the same groupId belong to the same Metavault. Therefore, to edit a Metavault, we only need to allow a call to deposit using the same groupId. On top of that, by changing the subgraph to use the name from the last DepositMinted, we allow editing the Metavault's name.