Audit [H-7] (#127)
This change addresses the unfair distribution of loss:
* prevent deposits when the total underlying is less than the total principal.
* prevent deposits to users who are already in debt.
* update logic to calculate how many shares a deposit is worth during a withdrawal.
* removes shares from the deposit because this number is no longer needed with the update logic.
* add tests provided by the auditors.

A user is considered in debt when the vault loses funds after she claims the yield. At this point, this user will have to wait for the vault to generate the lost yield and some more before she can claim again. During this process, any deposit to this user would lose money immediately, and because of that, we need to prevent deposits.