Emit TreasuryUpdated in Vault's Constructor (#143)
In the vault contract, the treasury always has a valid address, but this is not true in the subgraph. The treasury is always null in the subgraph until the `setTreasury` vault function is called and consequently emits the `TreasuryUpdated` event.
This PR emits the same event in the vault constructor to circumvent this problem.
Besides, this PR removes some useless docker deployment info that should not be in git.